**Foundation Work** - Extended the workspace membership to include the new composition crate and four extension crates so they build alongside existing services (`Cargo.toml:3`). - Replaced the in-tree Omni implementation with a dedicated extension crate, re-exported via services to keep existing call-sites working (`crates/services/src/services/omni/mod.rs:1`, `forge-extensions/omni/src/service.rs:1`). - Scaffolded lightweight crates for branch templates, config v7, and Genie integrations to receive future extractions (`forge-extensions/branch-templates/src/lib.rs:1`, `forge-extensions/config/src/v7.rs:1`, `forge-extensions/genie/src/lib.rs:1`). - Bootstrapped the `forge-app` composition layer with initial dependencies, service wiring, and a placeholder router entry point (`forge-app/Cargo.toml:1`, `forge-app/src/main.rs:1`, `forge-app/src/services/mod.rs:1`, `forge-app/src/router.rs:1`). - Added initial auxiliary migrations for forge-specific tables and compatibility views under the new app (`forge-app/migrations/001_auxiliary_tables.sql:1`, `forge-app/migrations/002_migrate_data.sql:1`). - Documented override strategy (`forge-overrides/README.md:1`) and reserved the upstream checkout location with setup instructions plus a `.gitmodules` stub (`upstream/README.md:1`, `.gitmodules:1`).
**Submodule Status**
- `git submodule add` was blocked by sandbox permissions; recorded the intended upstream mapping in `.gitmodules` and noted the follow-up command sequence in `upstream/README.md`.

**Testing**
- `cargo check -p forge-app`

**Gaps / Follow-Up**
1. Move branch template handling, config v7 logic, and Genie integrations out of the legacy crates into their new extension crates and update call-sites.
2. Replace the placeholder composition wiring with real upstream service composition once the submodule can be initialized.
3. Review the new migrations for compatibility with the production database backend (SQLite vs JSONB types) and align naming with upstream conventions.
