+++ Cargo.toml
@@ -1,6 +1,19 @@
 [workspace]
 resolver = "2"
-members = ["crates/server", "crates/db", "crates/executors", "crates/services", "crates/utils", "crates/local-deployment", "crates/deployment"]
+members = [
+    "crates/server",
+    "crates/db",
+    "crates/executors",
+    "crates/services",
+    "crates/utils",
+    "crates/local-deployment",
+    "crates/deployment",
+    "forge-app",
+    "forge-extensions/omni",
+    "forge-extensions/branch-templates",
+    "forge-extensions/config",
+    "forge-extensions/genie",
+]
 
 [workspace.dependencies]
 tokio = { version = "1.0", features = ["full"] }
