+++ forge-extensions/branch-templates/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "forge-extensions-branch-templates"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+anyhow = { workspace = true }
+sqlx = { version = "0.8.6", features = ["runtime-tokio-rustls", "sqlite", "chrono", "uuid"] }
+chrono = { version = "0.4", features = ["serde"] }
+uuid = { version = "1.0", features = ["v4", "serde"] }
+serde = { workspace = true }
+serde_json = { workspace = true }
