+++ forge-extensions/branch-templates/src/lib.rs
@@ -1,0 +1,7 @@
+pub struct BranchTemplateService;
+
+impl BranchTemplateService {
+    pub fn new() -> Self {
+        Self
+    }
+}
