+++ forge-extensions/config/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "forge-extensions-config"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+serde = { workspace = true }
+ts-rs = { workspace = true }
+anyhow = { workspace = true }
