+++ forge-extensions/config/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod v7;
+
+pub use v7::ForgeConfigV7;
