+++ forge-extensions/config/src/v7.rs
@@ -1,0 +1,8 @@
+use serde::{Deserialize, Serialize};
+use ts_rs::TS;
+
+#[derive(Clone, Debug, Default, Serialize, Deserialize, TS)]
+#[ts(export)]
+pub struct ForgeConfigV7 {
+    pub branch_templates_enabled: bool,
+}
