+++ forge-extensions/genie/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "forge-extensions-genie"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+anyhow = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+ts-rs = { workspace = true }
