+++ forge-extensions/genie/src/lib.rs
@@ -1,0 +1,7 @@
+pub struct GenieIntegration;
+
+impl GenieIntegration {
+    pub fn new() -> Self {
+        Self
+    }
+}
