+++ forge-extensions/omni/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "forge-extensions-omni"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+anyhow = { workspace = true }
+serde = { workspace = true }
+ts-rs = { workspace = true }
+reqwest = { version = "0.12", features = ["json"] }
+
+[dev-dependencies]
+serde_json = { workspace = true }
