+++ forge-extensions/omni/src/lib.rs
@@ -1,0 +1,6 @@
+pub mod client;
+pub mod types;
+
+mod service;
+
+pub use service::OmniService;
