+++ forge-overrides/README.md
@@ -1,0 +1,3 @@
+# Forge Overrides
+
+Use this directory to stage temporary patches that wrap or replace upstream behavior without forking the upstream sources. Any file placed here should be clearly documented and scheduled for removal once upstream offers native extension points.
